+++ src/huc/data.h
@@ -7,38 +7,38 @@
 
 extern struct const_array *const_ptr;
 extern struct const_array  const_var[MAX_CONST];
-extern int  const_val[MAX_CONST_VALUE];
+extern long  const_val[MAX_CONST_VALUE];
 extern char const_data[MAX_CONST_DATA];
-extern int  const_val_start;
-extern int  const_val_idx;
-extern int  const_data_start;
-extern int  const_data_idx;
-extern int  const_size;
-extern int  const_nb;
+extern long  const_val_start;
+extern long  const_val_idx;
+extern long  const_data_start;
+extern long  const_data_idx;
+extern long  const_size;
+extern long  const_nb;
 
 /* storage words */
 
 extern	char  symtab[];
 extern	char *glbptr, *rglbptr, *locptr;
-extern	int   ws[];
-extern	int  *wsptr;
-extern	int   swstcase[];
-extern	int   swstlab[];
-extern	int   swstp;
+extern	long   ws[];
+extern	long  *wsptr;
+extern	long   swstcase[];
+extern	long   swstlab[];
+extern	long   swstp;
 extern	char  litq[];
 extern	char  litq2[];
-extern	int   litptr;
+extern	long   litptr;
 extern	char  macq[];
-extern	int   macptr;
+extern	long   macptr;
 extern	char  line[];
 extern	char  mline[];
-extern	int   lptr, mptr;
+extern	long   lptr, mptr;
 
 extern  char  asmdefs[];
 
 /* miscellaneous storage */
 
-extern	int	nxtlab,
+extern	long	nxtlab,
 		litlab,
 		stkp,
 		zpstkp,
@@ -58,21 +58,21 @@ extern	FILE	*input, *input2, *output;
 extern	FILE	*inclstk[];
 
 extern char    inclstk_name[INCLSIZ][20];
-extern int     inclstk_line[];
+extern long     inclstk_line[];
 extern char    fname_copy[20];
-extern int     line_number;
+extern long     line_number;
 
-extern	int		inclsp;
+extern	long		inclsp;
 extern	char	fname[];
 
 extern	char	quote[];
 extern	char	*cptr;
-extern	int	*iptr;
-extern	int	fexitlab;
-extern	int	iflevel, skiplevel;
-extern	int	errfile;
-extern	int	sflag;
-extern	int	cdflag;
-extern	int	verboseflag;
-extern	int	startup_incl;
-extern	int	errs;
+extern	long	*iptr;
+extern	long	fexitlab;
+extern	long	iflevel, skiplevel;
+extern	long	errfile;
+extern	long	sflag;
+extern	long	cdflag;
+extern	long	verboseflag;
+extern	long	startup_incl;
+extern	long	errs;
